ChromeOS向けAndroid APKビルドを追加
