+++ Cargo.toml
@@ -3,6 +3,9 @@ name = "keystone-cc"
 version = "0.1.0"
 edition = "2024"
 
+[lib]
+crate-type = ["cdylib", "staticlib", "rlib"]
+
 [dependencies]
 bevy = { version = "0.18.1", features = ["wav"]}
 bevy_ecs = "0.18.1"
