+++ README.md
@@ -24,6 +24,24 @@ Notes:
 - Transparent window behavior is primarily tuned for macOS/Linux. On Windows, transparency or reveal effects may differ.
 - Typing prints characters; Backspace and Enter are handled. Window resize adjusts layout.
 
+## ChromeOS (Android APK)
+
+`scripts/build-chromeos-apk` creates a debug-signed universal APK for Android-enabled Chromebooks. It uses Docker because the Android SDK/NDK command-line tools are x86_64 binaries; on this ARM64 build host Docker runs them through binfmt.
+
+```sh
+./scripts/build-chromeos-apk
+```
+
+The resulting files are `dist/chromeos/keystone-cc-debug.apk` and its SHA-256 checksum. The APK contains both `arm64-v8a` and `x86_64` native libraries and requires Android 12 (API 31) or newer.
+
+To install it on a Chromebook, enable Android app development / ADB debugging in ChromeOS settings, connect with `adb`, then run:
+
+```sh
+adb install -r dist/chromeos/keystone-cc-debug.apk
+```
+
+Alternatively, transfer the APK to the Chromebook and open it with the Android package installer. This is a debug build only; it is not Play Store signed.
+
 ## Code License
 
 - Scope: All source code outside of the `assets/` directory (`src/`, `scripts/`, build files, documentation, etc.)
