+++ android/app/proguard-rules.pro
@@ -1,0 +1,1 @@
+# Debug APKs are not minified. Keep this file for the standard Android release task.
