+++ android/app/src/main/cpp/dummy.cpp
@@ -1,0 +1,3 @@
+extern "C" int keystone_cc_android_dummy() {
+    return 0;
+}
