+++ android/build.gradle
@@ -1,0 +1,3 @@
+plugins {
+    id "com.android.application" version "8.6.1" apply false
+}
